Do the ssl/openssl directory/link dance in both places it could occur, /usr/include and /usr/libdata/perl5/site_perl/*-openbsd.
Allows both base34 and comp34 to be upgraded from a stock install.

ok deraadt@.

+++ distrib/miniroot/upgrade.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-#	$OpenBSD: upgrade.sh,v 1.52 2003/08/17 21:11:35 krw Exp $
+#	$OpenBSD: upgrade.sh,v 1.53 2003/09/06 23:02:55 krw Exp $
 #	$NetBSD: upgrade.sh,v 1.2.4.5 1996/08/27 18:15:08 gwr Exp $
 #
 # Copyright (c) 1997-2002 Todd Miller, Theo de Raadt, Ken Westerback
@@ -123,17 +123,21 @@ if ! umount /mnt; then
 fi
 mount_fs
 
-# Prior to 3.4, /usr/include/ssl was a directory and /usr/include/openssl was a
-# link to it. With 3.4, /usr/include/openssl is the directory and
-# /usr/include/ssl is the link. This change causes the upgrade of comp34 to
-# fail, so adjust a normal pre-3.4 setup to the new setup.
-( cd /mnt/usr/include
-if [[ -d ssl && -L openssl ]]; then
-	rm openssl
-	mv ssl openssl
-	ln -s openssl ssl
-fi
-)
+# Prior to 3.4, ssl was a directory and openssl was a link to it, both in
+# /usr/include and /usr/libdata/perl5/site_perl/${ARCH}-openbsd. With 3.4,
+# openssl is the directory and ssl is the link. This change causes the upgrade
+# of base34 and comp34 to fail, so adjust a normal pre-3.4 setup to the new
+# setup.
+for _dir in /usr/include /usr/libdata/perl5/site_perl/*-openbsd; do
+	[[ -d /mnt$_dir ]] || continue
+	( cd /mnt$_dir
+	if [[ -d ssl && -L openssl ]]; then
+		rm openssl
+		mv ssl openssl
+		ln -s openssl ssl
+	fi
+	)
+done
 
 # Install sets.
 install_sets
